Inter- and extrapolation functions for tfd-objects

These are the currently available evaluator-functions for tfd-objects, which control how the entries are inter-/extrapolated to previously unseen arg-values. They all are merely wrappers around zoo::na.fill(), zoo::na.approx(), etc... Note that these are not meant to be called directly -- they are internal functions used by evaluate.tfd() to do its thing.

The list:

approx_linear for linear interpolation without extrapolation (i.e., zoo::na.approx() with na.rm = FALSE) -- this is the default,

approx_spline for cubic spline interpolation, (i.e., zoo::na.spline() with na.rm = FALSE),

approx_none in order to not inter-/extrapolate ever (i.e., zoo::na.fill() with fill = NA)

approx_fill_extend for linear interpolation and constant extrapolation (i.e., zoo::na.fill() with fill = "extend")

approx_locf for "last observation carried forward" (i.e., zoo::na.locf() with na.rm = FALSE and

approx_nocb for "next observation carried backward" (i.e., zoo::na.locf() with na.rm = FALSE, fromLast = TRUE).

For implementing your own, see source code of tidyfun:::zoo_wrapper.

approx_linear(x, arg, evaluations)

approx_spline(x, arg, evaluations)

approx_none(x, arg, evaluations)

approx_fill_extend(x, arg, evaluations)

approx_locf(x, arg, evaluations)

approx_nocb(x, arg, evaluations)

Arguments

x
new arg values to approximate the function for

arg
the arg values of the evaluations

evaluations
the function's values at arg

See also