Evaluate tfs, both inside or outside a data.frame

Evaluate tfs, both inside or outside a data.frame

tf_evaluate(object, ..., arg)

# S3 method for tfd
tf_evaluate(object, ..., arg, evaluator = tf_evaluator(object))

# S3 method for tfb
tf_evaluate(object, ..., arg)

# S3 method for data.frame
tf_evaluate(object, ..., arg)

Arguments

object

a tf, or a data.frame-like object with tf columns.

...

optional: a selection of tf-columns. If empty, all tf-variables in the data frame are selected. You can supply bare variable names, select all variables between x and z with x:z, exclude y with -y. For more options, see the dplyr::select() documentation.

arg

optional evaluation grid (vector or list of vectors). Defaults to tf_arg(object).

evaluator

optional. The function to use for inter/extrapolating the tfd. Defaults to tf_evaluator(object). See e.g. tf_approx_linear() for details.

Value

For tf-objects, a list of numeric vectors containing the function evaluations. For dataframes, replaces tf-columns with list columns of smaller data.frames containing the functions' arguments (arg) and evaluations (value).

Details

The arg-argument of tf_evaluate.data.frame method can be a list of arg-vectors or -lists used as the arg argument for the tf_evaluate()-method for the respective tf-columns in object. ... is not used for a tf-object, but a second unnamed argument to these methods will be interpreted as arg.

See also

Examples

library(dplyr)
a <- tf_rgp(3, arg = 7)
b <- tf_rgp(3, arg = 7)
d <- tibble(a = a, b = b)
tf_evaluate(a) %>% str()
#> List of 3
#>  $ : num [1:7] 1.132 1.465 1.562 0.426 -0.454 ...
#>  $ : num [1:7] 0.655 1.636 1.237 0.699 0.736 ...
#>  $ : num [1:7] -1.923 -1.527 -1.354 -1.164 -0.274 ...
tf_evaluate(a, arg = .5) %>% str()
#> List of 3
#>  $ : num 0.426
#>  $ : num 0.699
#>  $ : num -1.16
tf_evaluate(d) %>% glimpse()
#> Rows: 3
#> Columns: 2
#> $ a <list> [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>]
#> $ b <list> [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>]
tf_evaluate(d, -b) %>% glimpse()
#> Rows: 3
#> Columns: 2
#> $ a <list> [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>]
#> $ b <tfd_reg> <0.6361040, -0.4522793, -0.9191737, -0.6312561, -0.6584796, -0.5146…
tf_evaluate(d, a) %>% glimpse() 
#> Rows: 3
#> Columns: 2
#> $ a <list> [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>], [<tbl_df[7 x 2]>]
#> $ b <tfd_reg> <0.6361040, -0.4522793, -0.9191737, -0.6312561, -0.6584796, -0.5146…
grid <- seq(0, 1, l = 11)
# <arg> must be specified as named argument!
tf_evaluate(d, arg = grid) %>% glimpse() 
#> Rows: 3
#> Columns: 2
#> $ a <list> [<tbl_df[11 x 2]>], [<tbl_df[11 x 2]>], [<tbl_df[11 x 2]>]
#> $ b <list> [<tbl_df[11 x 2]>], [<tbl_df[11 x 2]>], [<tbl_df[11 x 2]>]