Integrals and Anti-Derivatives of Functional Data

Integrals of tf-objects are computed by simple quadrature (trapezoid rule, specifically). By default the scalar definite integral \(\int^{upper}_{lower}f(s)ds\) is returned (option definite = TRUE), alternatively for definite = FALSE something like the anti-derivative on [lower, upper], e.g. a tfd or tfb object representing \(F(t) \approx \int^{t}_{lower}f(s)ds\), for \(t \in\)[lower, upper], is returned.

tf_integrate(f, lower, upper, ...)

# S3 method for function
tf_integrate(f, lower, upper, ...)

# S3 method for tfd
tf_integrate(f, lower = tf_domain(f)[1],
  upper = tf_domain(f)[2], definite = TRUE, arg, ...)

# S3 method for tfb
tf_integrate(f, lower = tf_domain(f)[1],
  upper = tf_domain(f)[2], definite = TRUE, arg, ...)

Arguments

f
a tf-object

lower
lower limits of the integration range. For definite=TRUE, this can be a vector of the same length as f.

upper
upper limits of the integration range (but see definite arg / Description). For definite=TRUE, this can be a vector of the same length as f.

...
not used

definite
should the definite integral be returned (default) or the antiderivative. See Description.

arg
grid to use for the quadrature.

Value

For definite = TRUE, the definite integrals of the functions in f. For definite = FALSE and tf-inputs, a tf object containing their anti-derivatives

Methods (by class)

function: integrating R-functions (a wrapper for stats::integrate())

tfd: integrating tfd() objects

tfb: integrating tfd() objects