Gaussian Process random generator

The function generates n realizations of a Gaussian process, either with squared exponential covariance \(Cov(x(t), x(t')) = \exp(-(t'-t)^2)/s) + n \delta_{t}(t')\) or Wiener process covariance \(Cov(x(t), x(t')) = \min(t',t)/s + n \delta_{t}(t')\) with scale parameter s and nugget effect n.

tf_rgp(n, arg = 51L, scale = diff(range(arg))/10,
  cor = c("squareexp", "wiener"), nugget = scale/200)

Arguments

n
how many realizations to draw

arg
vector of evaluation points (arg of the return object). Defaults to (0, 0.02, 0.04, ..., 1). If a single integer (don't forget the L...), creates a grid of the given length over (0,1).

scale
scale parameter (see Description). Defaults to the width of the domain divided by 10.

cor
type of correlation structure to use. Currently available: "squareexp" or "wiener", see Description.

nugget
nugget effect for additional white noise / unstructured variability. Defaults to scale/200 (so: very little noise).

Value

an tfd-vector of length n