Constructors/converters for functional data in FPC-basis representation

These functions perform a (functional) principal component analysis of the input data and return an tfb tf-object that uses the empirical eigenfunctions as basis functions for representing the data. By default, a smoothed FPCA via refund::fpca.sc() is used to compute eigenfunctions and scores based on the smoothed empirical covariance. If smooth =FALSE, a fast, unregularized PCA of the data is done instead.

tfb_fpc(data, ...)

# S3 method for data.frame
tfb_fpc(data, id = 1, arg = 2, value = 3,
  domain = NULL, smooth = TRUE, resolution = NULL, ...)

# S3 method for matrix
tfb_fpc(data, arg = NULL, domain = NULL, smooth = TRUE,
  resolution = NULL, ...)

# S3 method for numeric
tfb_fpc(data, arg = NULL, domain = NULL, smooth = TRUE,
  resolution = NULL, ...)

# S3 method for tf
tfb_fpc(data, arg = NULL, smooth = TRUE, ...)

Arguments

data
a matrix, data.frame or list of suitable shape, or another tf-object.

...
arguments to the call to refund::fpca.sc() that computes the (regularized/smoothed) FPCA. Unless set by the user tidyfun uses pve = .995 to determine the truncation levels and uses bs = 15 basis functions for the mean function and the marginal bases for the covariance surface.

id
The name/number of the column defining which data belong to which function.

arg
numeric, or list of numerics. The evaluation grid. See Details. For the data.frame-method: the name/number of the column defining the evaluation grid.

value
The name/number of the column containing the function evaluations.

domain
range of the arg.

smooth
use smoothed mean function and smoothed covariance surface estimates from refund::fpca.sc() or simply perform a (truncated) PCA of the data? See Details.

resolution
resolution of the evaluation grid. See details.

Details

ATM, the unsmoothed version does not use orthogonal basis "functions" for irregular / non-equidistant grids and does not work for incomplete/irregular data.

See also