tidyfun: Tidy Functional Data Wrangling and Visualization

tidyfun makes data wrangling and exploratory analysis for functional data in R easier.

The tf package defines the core functional data classes and methods. tidyfun builds on those objects with:

data-frame workflows for functional data: tf_gather() etc.

graphics helpers for tf objects and tidy functional data frames: geom_spaghetti(), gglasagna()

documentation and examples that show how to use tf objects in tidy workflows

Start with vignette(package = "tidyfun") for the article index.

See also

Author

Maintainer: Fabian Scheipl fabian.scheipl@googlemail.com (ORCID) [copyright holder]

Authors:

Jeff Goldsmith

Julia Wrobel (ORCID)

Other contributors:

Maximilian Mücke (ORCID) [contributor]