Lasagna plots for tfs using ggplot2

Lasagna plots show one color bar for each function.

gglasagna(data, y, order = NULL, label = NULL, arg = NULL,
  order_by = NULL, order_ticks = TRUE, order_ticks_color = "black",
  order_ticks_linetype = 2, order_ticks_alpha = 0.5)

Arguments

data
the data. Duhh.

y
bare name of the tf column to visualize

order
(optional) bare name of a column in data to define vertical order of lasagna layers.

label
(optional) bare name of a column in data to define labels for lasagna layers. Defaults to names of y, if present, or row numbers.

arg
arg to evaluate y om

order_by
a function applied to each row in y[, arg] that must return a scalar value to define the order of lasagna layers.

order_ticks
add horizontal lines indicating borders between levels of order (if discrete)? Defaults to true.

order_ticks_color
defaults to "black"

order_ticks_linetype
defaults to 2

order_ticks_alpha
defaults to 0.5

Value

a `ggplot2`` object

Details

The vertical order of the lasagna layers is increasing in

order (if provided),

the values returned by order_by (if provided),

and the row number of the observations.

i.e., lowest values are on top so that by default the first layer is the first observation in data and the vertical order of the layers corresponds to the ordering of observations for arrange(data, order, order_by(value), row_number()).

Examples

# NOT RUN {
set.seed(1221)
data = expand.grid(group = 1:5, rep = 1:10)
data = dplyr::mutate(data, id = paste(group, rep, sep = "-"))
data$f = rgp(50)
data$fi = sparsify(data$f)
data$fb = tfb(data$f)
gglasagna(data, f, label = id)
gglasagna(data, f, label = id, order = group)
gglasagna(data, f, label = id, order = group)
gglasagna(data, f, label = id, order = depth(f))
gglasagna(data, f, label = id, order_by = first) +
  facet_wrap(~group, scales = "free")
# }