Evaluate tfs inside a data.frame

Evaluate tfs inside a data.frame

# S3 method for data.frame
tf_evaluate(object, ..., arg)

Arguments

object

a data.frame-like object with tf columns.

...

optional: a selection of tf-columns. If empty, all tf-variables in the data frame are selected. You can supply bare variable names, select all variables between x and z with x:z, exclude y with -y. For more options, see the dplyr::select() documentation.

arg

optional evaluation grid (vector or list of vectors). Defaults to tf_arg(object).

Value

Replaces tf-columns with list columns of smaller data.frames containing the functions' arguments (arg) and evaluations (value) and returns the modified nested dataframe.

Details

The arg-argument of tf_evaluate.data.frame method can be a list of arg-vectors or -lists used as the arg argument for the tf::tf_evaluate()-method for the respective tf-columns in object. ... is not used for a tf-object, but a second unnamed argument to these methods will be interpreted as arg.

See also