Spread a tf-column into many columns representing the function evaluations.

Similar in spirit to tidyr::spread(), but does NOT shorten, just widens the data frame -- a tf-column is spread out into many columns containing the functional measurements.

tf_spread(data, value, arg, sep = "_", interpolate = TRUE)

Arguments

data
a data frame with at least one tf-column

value
the name of the tf-column to 'spread'/evaluate. You can supply bare variable names etc., see the dplyr::select() documentation. Also works without this if there's only one tf in data, see examples.

arg
(Semi-)optional. A vector of arg-values on which to evaluate the functions. If not provided, uses the default args. Needs to be specified for f_irreg.

sep
separating character used to create column names for the new columns, defaults to "_" for column names "<name of the tf>_<arg-value>". Set to NULL to get column names that only contain the arg-value.

interpolate
interpolate-argument for evaluating the functional data. Defaults to TRUE, i.e., tfds are inter/extrapolated on unobserved arg-values.

Examples

d <- tibble::tibble(g = 1:3)
d$f <- tf_rgp(3, 11L)
tf_spread(d, f)

#> # A tibble: 3 x 12
#>       g    f_0  f_0.1   f_0.2 f_0.3   f_0.4  f_0.5   f_0.6 f_0.7 f_0.8   f_0.9
#>   <int>  <dbl>  <dbl>   <dbl> <dbl>   <dbl>  <dbl>   <dbl> <dbl> <dbl>   <dbl>
#> 1     1 -1.30  -0.817 -0.163  0.102 -0.0153 -0.237 -0.0631 0.412 0.809  1.04  
#> 2     2 -1.10  -0.678  0.0225 0.764  1.11    1.31   1.46   1.30  0.751  0.0323
#> 3     3 -0.428 -0.299  0.0259 0.396  0.676   0.910  1.04   0.913 0.440 -0.271 
#> # … with 1 more variable: f_1 <dbl>

tf_spread(d, -g)

#> # A tibble: 3 x 12
#>       g    f_0  f_0.1   f_0.2 f_0.3   f_0.4  f_0.5   f_0.6 f_0.7 f_0.8   f_0.9
#>   <int>  <dbl>  <dbl>   <dbl> <dbl>   <dbl>  <dbl>   <dbl> <dbl> <dbl>   <dbl>
#> 1     1 -1.30  -0.817 -0.163  0.102 -0.0153 -0.237 -0.0631 0.412 0.809  1.04  
#> 2     2 -1.10  -0.678  0.0225 0.764  1.11    1.31   1.46   1.30  0.751  0.0323
#> 3     3 -0.428 -0.299  0.0259 0.396  0.676   0.910  1.04   0.913 0.440 -0.271 
#> # … with 1 more variable: f_1 <dbl>

tf_spread(d)

#> # A tibble: 3 x 12
#>       g    f_0  f_0.1   f_0.2 f_0.3   f_0.4  f_0.5   f_0.6 f_0.7 f_0.8   f_0.9
#>   <int>  <dbl>  <dbl>   <dbl> <dbl>   <dbl>  <dbl>   <dbl> <dbl> <dbl>   <dbl>
#> 1     1 -1.30  -0.817 -0.163  0.102 -0.0153 -0.237 -0.0631 0.412 0.809  1.04  
#> 2     2 -1.10  -0.678  0.0225 0.764  1.11    1.31   1.46   1.30  0.751  0.0323
#> 3     3 -0.428 -0.299  0.0259 0.396  0.676   0.910  1.04   0.913 0.440 -0.271 
#> # … with 1 more variable: f_1 <dbl>