Spread a tf-column into many columns representing the function evaluations.

Similar in spirit to tidyr::spread(), but does NOT shorten, just widens the data frame -- a tf-column is spread out into many columns containing the functional measurements.

tf_spread(data, value, arg, sep = "_", interpolate = FALSE)

Arguments

data

a data frame with at least one tf-column

value

the name of the tf-column to 'spread'/evaluate. You can supply bare variable names etc., see the dplyr::select() documentation. Also works without this if there's only one tf in data, see examples.

arg

(Semi-)optional. A vector of arg-values on which to evaluate the functions. If not provided, uses the default args. Must be specified for tf_irreg.

sep

separating character used to create column names for the new columns, defaults to "_" for column names "<name of the tf>_<arg-value>". Set to NULL to get column names that only contain the arg-value.

interpolate

interpolate-argument for evaluating the functional data. Defaults to FALSE, i.e., tfds are not inter/extrapolated on unobserved arg-values.

Examples

d <- dplyr::tibble(g = 1:3)
d$f <- tf_rgp(3, 11L)
tf_spread(d, f)
#> # A tibble: 3 × 12
#>       g    f_0  f_0.1  f_0.2  f_0.3  f_0.4  f_0.5  f_0.6  f_0.7   f_0.8  f_0.9
#>   <int>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#> 1     1  0.236 -0.235 -0.920 -1.44  -1.68  -1.81  -1.82  -1.27   0.0419  1.72 
#> 2     2 -1.18  -1.65  -2.07  -2.08  -1.71  -1.25  -0.828 -0.360  0.140   0.785
#> 3     3 -0.133 -0.579 -0.597 -0.425 -0.445 -0.379 -0.196 -0.180 -0.530  -1.03 
#> # … with 1 more variable: f_1 <dbl>
tf_spread(d, -g)
#> # A tibble: 3 × 12
#>       g    f_0  f_0.1  f_0.2  f_0.3  f_0.4  f_0.5  f_0.6  f_0.7   f_0.8  f_0.9
#>   <int>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#> 1     1  0.236 -0.235 -0.920 -1.44  -1.68  -1.81  -1.82  -1.27   0.0419  1.72 
#> 2     2 -1.18  -1.65  -2.07  -2.08  -1.71  -1.25  -0.828 -0.360  0.140   0.785
#> 3     3 -0.133 -0.579 -0.597 -0.425 -0.445 -0.379 -0.196 -0.180 -0.530  -1.03 
#> # … with 1 more variable: f_1 <dbl>
tf_spread(d)
#> # A tibble: 3 × 12
#>       g    f_0  f_0.1  f_0.2  f_0.3  f_0.4  f_0.5  f_0.6  f_0.7   f_0.8  f_0.9
#>   <int>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#> 1     1  0.236 -0.235 -0.920 -1.44  -1.68  -1.81  -1.82  -1.27   0.0419  1.72 
#> 2     2 -1.18  -1.65  -2.07  -2.08  -1.71  -1.25  -0.828 -0.360  0.140   0.785
#> 3     3 -0.133 -0.579 -0.597 -0.425 -0.445 -0.379 -0.196 -0.180 -0.530  -1.03 
#> # … with 1 more variable: f_1 <dbl>