Constructors & convertors for functional data in (spline) basis representation

Various constructor and conversion methods.

as.tfb(data, basis = c("mgcv", "fpc"), ...)

# S3 method for tfb
as.data.frame(x, row.names = NULL, optional = FALSE,
  arg = NULL, ...)

# S3 method for tfb
as.matrix(x, arg = NULL, ...)

tfb(data, ...)

# S3 method for data.frame
tfb(data, id = 1, arg = 2, value = 3,
  domain = NULL, penalized = TRUE, resolution = NULL, ...)

# S3 method for matrix
tfb(data, arg = NULL, domain = NULL,
  penalized = TRUE, resolution = NULL, ...)

# S3 method for numeric
tfb(data, arg = NULL, domain = NULL,
  penalized = TRUE, resolution = NULL, ...)

# S3 method for list
tfb(data, arg = NULL, domain = NULL, penalized = TRUE,
  resolution = NULL, ...)

# S3 method for tfd
tfb(data, arg = NULL, domain = NULL, penalized = TRUE,
  resolution = NULL, ...)

# S3 method for tfb
tfb(data, arg = NULL, domain = NULL, penalized = TRUE,
  resolution = NULL, ...)

Arguments

data
a matrix, data.frame or list of suitable shape, or another tf-object.

basis
either "mgcv" to call tfb() which uses mgcv-type spline basis functions or "fpc" to call tfb_fpc() which uses a (smoothed) functional principal component basis.

...
for tfb: arguments to the calls to mgcv::s() setting up the basis and mgcv::magic() (if penalized is TRUE). If not user-specified here, tidyfun uses k=15 cubic regression spline basis functions (i.e., bs = "cr") by default, but at least how many basis functions k the spline basis should have probably needs to be set manually.
for as.tfb: use this to give arguments to tfb() or tfb_fpc().

x
an tfb object

row.names
not used

optional
not used

arg
numeric, or list of numerics. The evaluation grid. See Details. For the data.frame-method: the name/number of the column defining the evaluation grid.

id
The name/number of the column defining which data belong to which function.

value
The name/number of the column containing the function evaluations.

domain
range of the arg.

penalized
should the coefficients of the basis representation be estimated via mgcv::magic() (default) or ordinary least squares.

resolution
resolution of the evaluation grid. See details for tfd().

Value

an tfb-object (or a data.frame/matrix for the conversion functions, obviously.)

Details

tfb tries to represent the input data as linear combinations of a set of common spline basis functions identical for all observations and coefficient vectors estimated for each observation. The basis used is set up via a call to mgcv::s() and all the spline bases discussed in mgcv::smooth.terms are available, in principle. Depending on the value of the penalized-flag, the coefficient vectors for each observation are then estimated via fitting a (small) GAM for each observation via mgcv::magic() or via simple ordinary least squares.

After the "smoothed" representation is computed, the amount of smoothing that was performed is reported in terms of the "percentage of variance preserved", which is the variance of the smoothed function values divided by the variance of the original values. The ... arguments supplies arguments to both the spline basis set up (via mgcv::s()) and the estimation (via mgcv::magic()), most important: how many basis functions k the spline basis should have, the default is 25.

See tfb_fpc() for using an FPC representation with an orthogonal basis estimated from the data instead.

See also