Functions that summarize tf objects

These will return a tf object containing the respective functional statistic. summary returns a vector with the mean function, the variance function, and the functional range of the central half of the functions, as defined by the functional

# S3 method for tf
mean(x, ...)

# S3 method for tf
median(x, na.rm = FALSE, depth = c("MBD", "pointwise"), ...)

# S3 method for tf
quantile(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = TRUE,
  type = 7,
  ...
)

sd(x, na.rm = FALSE)

# S3 method for default
sd(x, na.rm = FALSE)

# S3 method for tf
sd(x, na.rm = FALSE)

var(x, y = NULL, na.rm = FALSE, use)

# S3 method for default
var(x, y = NULL, na.rm = FALSE, use)

# S3 method for tf
var(x, y = NULL, na.rm = FALSE, use)

# S3 method for tf
summary(object, ...)

Arguments

x
a tf object functions, see source code.

...
further arguments passed to or from other methods.

na.rm
logical; if true, any NA and NaN's are removed from x before the quantiles are computed.

depth
method used to determine the most central element in x, i.e., the median. One of the functional data depths available via depth() or "pointwise" for a pointwise median function.

probs
numeric vector of probabilities with values in \([0,1]\). (Values up to 2e-14 outside that range are accepted and moved to the nearby endpoint.)

names
logical; if true, the result has a names attribute. Set to FALSE for speedup with many probs.

type
an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used.

y
NULL (default) or a vector, matrix or data frame with compatible dimensions to x. The default is equivalent to y = x (but more efficient).

use
an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".

object
a tfd object